Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/interface-requirements-sw.https-expected.txt


PASS Interfaces and attributes in ServiceWorkerGlobalScope
PASS Event constructors
PASS xhr is not exposed
FAIL URL.createObjectURL is not exposed assert_false: expected false got true
